Home Work H13

+++ src/ru/progwards/java1/lessons/compare/Divisibility.java
@@ -8,12 +8,8 @@ public class Divisibility {
             System.out.println("Остаток от деления " + num1 + " на " + num2 + " равен " + (num1%num2));
 
         if (num1/num2!=0)
-            System.out.println("Частное от деления " + num1 + " на " + num2 + " равно " + num1/num2);
+            System.out.println("Частное от деления " + num1 + " на " + num2 + " равно " +(double)num1/num2);
         else
             System.out.println("На ноль делить нельзя");
     }
-
-
-
-
 }
